docshell/test/test_bug580069.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=580069
-->
<head>
  <title>Test for Bug 580069</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=580069">Mozilla Bug 580069</a>

<iframe id='iframe' src='file_bug580069_1.html'></iframe>

<script type="application/javascript">

SimpleTest.waitForExplicitFinish();

var iframe = document.getElementById('iframe');
var iframeCw = iframe.contentWindow;

// Called when file_bug580069_1.html loads.
function page1Load() {
  // This should cause us to load file 2.
  dump('page1Load\n');
  iframeCw.document.getElementById('form').submit();
}

// Called when file_bug580069_2.html loads.
var page2Loads = 0;
function page2Load(method) {

  dump("iframe's location is: " + iframeCw.location + ", method is " + method + "\n");

  if (page2Loads == 0) {
    is(method, "POST", "Method for first load should be POST.");
    iframeCw.history.replaceState('', '', '?replaced');

    // This refresh shouldn't pop up the "are you sure you want to refresh a page
    // with POST data?" dialog.  If it does, this test will hang and fail, and
    // we'll see 'Refreshing iframe...' at the end of the test log.
    dump('Refreshing iframe...\n');
    iframeCw.location.reload();
  }
  else if (page2Loads == 1) {
    is(method, "GET", "Method for second load should be GET.");
    is(iframeCw.location.search, "?replaced", "Wrong search on iframe after refresh.");
    SimpleTest.finish();
  }
  else {
    ok(false, "page2Load should only be called twice.");
  }

  page2Loads++;
}
</script>

</body>
</html>
